refactor: rename file form keys & delete validate map
