+++ src/const/form-keys.ts
@@ -1,0 +1,9 @@
+export enum FormKeys {
+  TITLE = 'title',
+  DATE = 'date',
+  FILE = 'file',
+  SELECT = 'select',
+  CHECK = 'check',
+  RADIO = 'radio',
+  DESCRIPTIONS = 'description',
+}
