+++ src/const/index.ts
@@ -4,5 +4,5 @@ export * from './local-storage';
 export * from './menu-items';
 export * from './select-options';
 export * from './title-forms';
-export * from './validate-form-keys';
+export * from './form-keys';
 export * from './page-title';
